pyproject.toml
[project]
name = "docusense"
version = "0.1.0"
description = "DocuSense is a powerful PDF parsing and content extraction tool that leverages Large Language Models (LLMs) and Vision Language Models (VLMs) to extract, process, and summarize content from PDF documents. Whether your PDF contains text, images, or a combination of both, DocuSense provides a comprehensive solution for extracting meaningful information."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "fitz>=0.0.1.dev2",
    "numpy>=2.2.3",
    "openai>=1.66.2",
    "pillow>=11.1.0",
    "pymupdf>=1.25.3",
    "pytesseract>=0.3.13",
]
